Bug Fixed
Made images more responsive by applying bootstrap on the styling. Added
an Admin Login button that directs to the admin system. Fixed the error
page after clicking Complete Order - this was caused by using the
smtpClient.PickupDirectoryLocation function in my email order processor,
it only stores the email in a local folder which cannot be done once the
file is deployed and online. I left the processor in as a reminder for
future reference when it comes to actually applying for an order
processor, meanwhile I have commented the functionalities of the
processor so it does not interfere with the rest of the application.