Empty Cart
In the CartController I have added an If statement for if the product
count is 0 then display the error message "Cart is Empty" using the
Validaiton Summary in my Cart Index view. I have also used another If
statement to display the button that leads to the checkout page only if
the Model.Cart.Lines.Count() is not 0, otherwise the button remains
hidden.